extract restClientBuilder configuration to methods
